requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
SQLAlchemy==1.4.23
Flask-SQLAlchemy==2.5.1
Flask-JWT-Extended==4.3.1
Flask-Bcrypt==0.7.1
python-dotenv==0.19.0
Flask-Migrate==3.1.0
geopy==2.4.1
requests==2.31.0 